+++ src/stream_arbiter.sv
@@ -14,8 +14,8 @@
 // arbitration scheme is round-robin with "look ahead", see the `rrarbiter` for details.
 
 module stream_arbiter #(
-    parameter type      DATA_T = logic,  // Vivado requires a default value for type parameters.
-    parameter integer   N_INP
+    parameter type      DATA_T = logic,   // Vivado requires a default value for type parameters.
+    parameter integer   N_INP = -1        // Synopsys DC requires a default value for parameters.
 ) (
     input  logic              clk_i,
     input  logic              rst_ni,
@@ -29,27 +29,19 @@ module stream_arbiter #(
     input  logic              oup_ready_i
 );
 
-  logic [$clog2(N_INP)-1:0] idx;
-
-  rrarbiter #(
-    .NUM_REQ  (N_INP),
-    // Lock arbitration decision once the output is valid and until the handshake happens.
-    .LOCK_IN  (1)
-  ) i_arbiter (
-    .clk_i    (clk_i),
-    .rst_ni   (rst_ni),
-    .flush_i  (1'b0),
-    .en_i     (oup_ready_i),
-    .req_i    (inp_valid_i),
-    .ack_o    (inp_ready_o),
-    // The `vld_o` port of `rrarbiter` combinatorially depends on `en_i`.  In the stream protocol,
-    // a valid may not depend on a ready, so we drive `oup_valid_o` from the `inp_valid_i`s in (1)
-    // and leave `vld_o` unconnected.
-    .vld_o    (),
-    .idx_o    (idx)
+  stream_arbiter_flushable #(
+    .DATA_T (DATA_T),
+    .N_INP  (N_INP)
+  ) i_arb (
+    .clk_i        (clk_i),
+    .rst_ni       (rst_ni),
+    .flush_i      (1'b0),
+    .inp_data_i   (inp_data_i),
+    .inp_valid_i  (inp_valid_i),
+    .inp_ready_o  (inp_ready_o),
+    .oup_data_o   (oup_data_o),
+    .oup_valid_o  (oup_valid_o),
+    .oup_ready_i  (oup_ready_i)
   );
-
-  assign oup_valid_o = (|inp_valid_i); // (1), see reference above.
-  assign oup_data_o = inp_data_i[idx];
 
 endmodule